Исправлен и упрощен загрузчик css. Исправлены тесты, теперь они полностью запускается в phantomjs
