+++ test/sandbox-manager.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="utf-8">
+	<title>requrejs-sandbox grunt test suite</title>
+	<link rel="stylesheet" href="static/qunit.css">
+	<script src="static/qunit.js"></script>
+	<script type="text/javascript">
+		QUnit.config.autostart = false;
+	</script>
+	<script src="../static/js/libs/require.min.js"></script>
+	<script src="../build/requirejs-sandbox.js"></script>
+	<script src="sandbox-manager.test.js"></script>
+</head>
+<body>
+	<div id="qunit"></div>
+	<div id="qunit-fixture"></div>
+</body>
+</html>
