+++ test/transits.html
@@ -5,9 +5,12 @@
 	<title>requrejs-sandbox grunt test suite</title>
 	<link rel="stylesheet" href="static/qunit.css">
 	<script src="static/qunit.js"></script>
+	<script type="text/javascript">
+		QUnit.config.autostart = false;
+	</script>
 	<script src="../static/js/libs/require.min.js"></script>
 	<script src="../build/requirejs-sandbox.js"></script>
-	<script src="requirejs-sandbox.test.js"></script>
+	<script src="transits.test.js"></script>
 </head>
 <body>
 	<div id="qunit"></div>
